Initial check in of libocr-glibc.
This is a shim between glibc and ocr.  Included are changes to newlib
sys/ocr sources to allow for multi-use; i.e. xstg newlib, x86 C newlib,
and x86 C and C++ glibc support. Using weak_alias instead of #defines
to provide glibc compatible syscall names.

This implementation requires that the user application use mainEdt()
instead of main() as the starting point.  In this case mainEdt() has
the same return type and arguments as a normal main().

+++ apps/libs/src/newlib/newlib-2.1.0/newlib/libc/sys/ocr/ocr_shim.h
@@ -7,9 +7,9 @@
  * These methods support translating between the libc notions of open
  * file ids and the associated OCR Guids
  */
-short ocrLibAddGuid( struct _reent * r, ocrGuid_t guid );
-short ocrLibGuidToFd( struct _reent * r, ocrGuid_t guid );
-ocrGuid_t ocrLibRmFd( struct _reent * r, short fd );
-ocrGuid_t ocrLibFdToGuid( struct _reent * r, short fd );
+int ocrLibAddGuid( struct _reent * r, ocrGuid_t guid );
+int ocrLibGuidToFd( struct _reent * r, ocrGuid_t guid );
+ocrGuid_t ocrLibRmFd( struct _reent * r, int fd );
+ocrGuid_t ocrLibFdToGuid( struct _reent * r, int fd );
 
 #endif // _OCR_SHIM_
